Move pept2filtered to separate worker
