+++ src/composables/communication/unipept/usePept2filtered.ts
@@ -1,62 +1,51 @@
-import {ShareableMap} from "shared-memory-datastructures";
-import {parallelLimit} from "async";
-import {ref} from "vue";
+import {ShareableMap, TransferableState} from "shared-memory-datastructures";
+import {ref, shallowRef, toRaw} from "vue";
 import {DEFAULT_API_BASE_URL} from "@/logic/Constants";
 import PeptideData from "@/logic/ontology/peptides/PeptideData";
 import PeptideDataSerializer from "@/logic/ontology/peptides/PeptideDataSerializer";
-import {Filter, FilterType} from "@/store/CustomFilterStore";
+import {Filter} from "@/store/CustomFilterStore";
+import Pept2filteredWebWorker from "./pept2filtered.worker.ts?worker&inline";
+import useAsyncWebWorker from "@/composables/useAsyncWebWorker";
+
+export interface Pept2filteredData {
+    peptides: string[],
+    equate: boolean,
+    filter: Filter | undefined,
+    baseUrl: string,
+    batchSize: number,
+    parallelRequests: number
+}
+
+export interface Pept2filteredWorkerOutput {
+    peptToDataTransferable: TransferableState;
+}
 
 export default function usePept2filtered(
     baseUrl = DEFAULT_API_BASE_URL,
     batchSize = 400,
     parallelRequests = 5,
 ) {
-    const peptideData = ref<ShareableMap<string, PeptideData>>();
+    const peptideData = shallowRef<ShareableMap<string, PeptideData>>();
+
+    const { post } = useAsyncWebWorker<Pept2filteredData, Pept2filteredWorkerOutput>(
+        () => new Pept2filteredWebWorker()
+    );
 
     const process = async (
         peptides: string[],
         equate: boolean,
         filter: Filter | undefined
     ) => {
-        const result = new ShareableMap<string, PeptideData>({ serializer: new PeptideDataSerializer() });
+        const { peptToDataTransferable } = await post({
+            peptides,
+            equate,
+            filter: toRaw(filter),
+            baseUrl,
+            batchSize,
+            parallelRequests
+        });
 
-        // TODO: caching
-
-        const requests = [];
-        for(let i = 0; i < peptides.length; i += batchSize) {
-            requests.push(async () => {
-                const response = await fetch(`${baseUrl}/mpa/pept2data`, {
-                    method: "POST",
-                    body: JSON.stringify({
-                        peptides: peptides.slice(i, i + batchSize),
-                        equate_il: equate,
-                        report_taxa: true,
-                        ...constructFilterJson(filter)
-                    }),
-                    headers: { "Content-Type": "application/json" }
-                }).then(r => r.json());
-
-                for (const peptide of response.peptides) {
-                    result.set(peptide.sequence, PeptideData.createFromPeptideDataResponse(peptide));
-                }
-            });
-        }
-
-        await parallelLimit(requests, parallelRequests);
-
-        peptideData.value = result;
-    }
-
-    const constructFilterJson = (filter: Filter | undefined) => {
-        if (!filter) return {};
-
-        switch (filter.filter) {
-            case FilterType.Taxon: return { filter: { taxa: filter.data } };
-            case FilterType.Proteome: return { filter: { proteomes: filter.data } };
-            case FilterType.Protein: return { filter: { proteins: filter.data } };
-            case FilterType.UniProtKB:
-            default: return {}
-        }
+        peptideData.value = ShareableMap.fromTransferableState<string, PeptideData>(peptToDataTransferable, { serializer: new PeptideDataSerializer() });
     }
 
     return {
